+++ apps/web/src/modules/CourseSearch/hooks/useSearchCourseQueryParams/types.ts
@@ -3,6 +3,7 @@ import { StudyProgram } from '@cgr/codegen'
 export interface QueryParams {
   keyword?: string | null
   genEdTypes?: string
+  gradingTypes?: string
   dayOfWeeks?: string
   limit?: number
   offset?: number